refactor(shortcut): extract platform logic into internal/shortcut package
The shortcut generation logic (launch resolution, npx detection, platform
shortcut creation) is now in its own package with no CLI dependency, so a
future HTTP handler can reuse it. internal/command/shortcut.go keeps only
the Kong CLI shell and PersistLaunchSource (a CLI-boundary concern).
